# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Swagger OpenApi description",
        "contact": {
            "email": "akcaertugrulgazi@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/blogs": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "List blogs",
                "description": "List blogs",
                "operationId": "v1ListBlog",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Store a blog",
                "description": "Store a blog",
                "operationId": "v1StoreBlog",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Blog"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/blogs/{id}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Show a blog",
                "description": "Show a blog",
                "operationId": "v1ShowBlog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Blog id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Update a blog",
                "description": "Update a blog",
                "operationId": "v1UpdateBlog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Blog id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Blog"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Delete a blog",
                "description": "Delete a blog",
                "operationId": "v1DeleteBlog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Blog id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Blog": {
                "properties": {
                    "title": {
                        "description": "Blog title",
                        "type": "string"
                    },
                    "body": {
                        "description": "Blog body",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}